% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDOIF.R
\name{EDOIF}
\alias{EDOIF}
\title{Empirical Distribution Ordering Inference Framework (EDOIF)}
\usage{
EDOIF(Values, Group, bootT, alpha, methodType)
}
\arguments{
\item{Values}{is a vector of real-number values}

\item{Group}{is a vector of categories of each real number in Values}

\item{bootT}{is a number of times of sample with replacement for bootstraping.
The default is 1000.
 It must be above zero}

\item{alpha}{is a significance level using in both confidence intervals and ordering inference it has the range [0,1].
The default is 0.05.}

\item{methodType}{is an option for boostrapping methods:either  "perc" or "bca".
The "perc" is the default option.}
}
\value{
obj an object of EDOIF class that contains the results of ordering inference
that can be print in text mode (print(obj)) or graphic mode (plot(obj))
}
\description{
EDOIF is  non-parametric  framework  based on  Esitmation Statistics principle.
Its main purpose is to infer orders of empirical distributions from different categories
base on a probability of finding a value in one distribution that greater than the expectation
of another distribution.
}
\details{
Given a set of ordered-pair of real-category values the framework is capable of
 1) inferring orders of  domination  of  categories  and  representing  orders  in  the form of a graph;
 2) estimating  magnitude  of  difference  between  a  pair  of categories in forms of confidence intervals; and
 3) visualizing  domination  orders  and  magnitudes  of  dif-ference of categories.
}
