% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphOperation.R
\name{getADJNetDen}
\alias{getADJNetDen}
\title{getADJNetDen function}
\usage{
getADJNetDen(adjMat)
}
\arguments{
\item{adjMat}{is an adjacency matrix of a dominant-distribution network.}
}
\value{
This function returns a value of network density of of a dominant-distribution network for a given adjMat.
}
\description{
getADJNetDen is a support function for calculating a network density of a dominant-distribution network.
}
\examples{
# Generate simulation data with 100 samples per categories

simData<-SimNonNormalDist(nInv=100)

# Performing ordering infernce from simData

resultObj<-EDOIF(simData$Values,simData$Group)

# Get a network density of an adjacency matrix

getADJNetDen(adjMat=resultObj$adjMat)

}
