% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfInv.R
\name{getConfInv}
\alias{getConfInv}
\title{getConfInv function}
\usage{
getConfInv(Values, Group, GroupList, bootT, alpha, methodType)
}
\arguments{
\item{Values}{is a vector of real-number values}

\item{Group}{is a vector of categories of each real number in Values}

\item{GroupList}{is a list of names of categories ascendingly ordered by their means.}

\item{bootT}{is a number of times of sample with replacement for bootstraping.
The default is 1000. It must be above zero}

\item{alpha}{is a significance level using in both confidence intervals and ordering inference it has the range [0,1].
The default is 0.05.}

\item{methodType}{is an option for boostrapping methods:either  "perc" or "bca".
The "perc" is the default option.}
}
\value{
This function returns a list of mean confidence intervals.

\item{confInvsList[i,]}{ The mean confidence interval of sortedGroupList[i] category. confInvsList[i,1] is a lower bound and confInvsList[i,2] is an upper bound. }
}
\description{
getConfInv is a support function for bootstraping method.
Its main purpose is to compute a mean confidence intervals of all distributions.
}
