% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDominantNet.R
\name{getDominantRADJ}
\alias{getDominantRADJ}
\title{getDominantRADJ function}
\usage{
getDominantRADJ(MegDiffList, methodType)
}
\arguments{
\item{MegDiffList}{is a list of objects that contains mean-difference confidence intervals inferred by getMegDiffConfInv function.}

\item{methodType}{is an option for boostrapping methods:either  "perc" or "bca".}
}
\value{
This function returns an adjacency matrix of a dominant-distribution network \code{adjMat}
and the corresponding lower-bound of mean difference CIs \code{adjDiffMat}.

\item{adjDiffMat[i,j]}{ A lower bound of confidence interval of mean difference for j minus i using methodType bootstrap. }
\item{adjMat[i,j]}{ An element of adjacency matrix: One if adjDiffMat[i,j] is positive, otherwise, zero. }
}
\description{
getDominantRADJ is a support function for inferring a dominant-distribution network using mean-difference confidence intervals.
}
