% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDominantNet.R
\name{getttestDominantRADJ}
\alias{getttestDominantRADJ}
\title{getttestDominantRADJ function}
\usage{
getttestDominantRADJ(Values, Group, GroupList, alpha)
}
\arguments{
\item{Values}{is a vector of real-number values}

\item{Group}{is a vector of categories of each real number in Values}

\item{GroupList}{is a list of names of categories ascendingly ordered by their means.}

\item{alpha}{is a significance level using in both confidence intervals and ordering inference it has the range [0,1].}
}
\value{
This function returns an adjacency matrix of a dominant-distribution network \code{adjMat}
and the corresponding p-values of all category pairs.

\item{adjMat[i,j]}{ An element of adjacency matrix: one if GroupList[j] category dominates GroupList[i] using Student's t-test, otherwise zero. }
\item{pValMat[i,j]}{ A p-value of Student's t-test for adjMat[i,j]. }
}
\description{
getttestDominantRADJ is a support function for inferring a dominant-distribution network using Student's t-test.
}
