% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraphphics.R
\name{plotGraph}
\alias{plotGraph}
\title{plotGraph function}
\usage{
plotGraph(obj, rankFlag)
}
\arguments{
\item{obj}{is an object of EDOIF class that contains the results of ordering inference.}

\item{rankFlag}{is an option for including ranks of categories with in the plot: default is TRUE for including ranks.}
}
\value{
This function returns an object of iGraph for a dominant-distribution network and its plot variable.

\item{graphVar}{ An object of iGraph for a dominant-distribution network }
}
\description{
plotGraph is a support function for plotting a dominant-distribution network from an adjacency matrix.
}
