% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraphphics.R
\name{plotMeanDiffCIs}
\alias{plotMeanDiffCIs}
\title{plotMeanDiffCIs function}
\usage{
plotMeanDiffCIs(obj, NList, fontSize, rankFlag)
}
\arguments{
\item{obj}{is an object of EDOIF class that contains the results of ordering inference.}

\item{NList}{is a list of based categories users want to have in mean-difference CI plot.}

\item{fontSize}{is a font size of text for all plots.}

\item{rankFlag}{is an option for including ranks of categories with in the plot: default is TRUE for including ranks.}
}
\value{
This function returns an object of ggplot class.

\item{pDiffCI}{ An object of ggplot class containing the plot of difference-mean confidence intervals }
}
\description{
plotMeanDiffCIs is a support function for plotting difference-mean confidence intervals.
}
