% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotText.R
\name{print.EDOIF}
\alias{print.EDOIF}
\title{print.EDOIF function}
\usage{
\method{print}{EDOIF}(x, ...)
}
\arguments{
\item{x}{is an object of EDOIF class that contains the results of ordering inference.}

\item{...}{Signature for S3 generic function.}
}
\description{
print.EDOIF is a support function for printing results of ordering inference in text.
}
\examples{
# Generate simulation data with 100 samples per categories

simData<-SimNonNormalDist(nInv=100)

# Performing ordering infernce from simData

resultObj<-EDOIF(simData$Values,simData$Group)

# Print results in text mode

print(resultObj)


}
