% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraphphics.R
\name{printGraph}
\alias{printGraph}
\title{printGraph function}
\usage{
printGraph(sortedGroupList, confInvsList, sortedmeanList, adjMat)
}
\arguments{
\item{sortedGroupList}{is a list of names of categories ascendingly ordered by their means.}

\item{confInvsList}{is a list of mean confidence intervals of all categories.}

\item{sortedmeanList}{is a list of means of categories that are ascendingly ordered.}

\item{adjMat}{is an adjacency matrix of dominant-distribution network}
}
\description{
printGraph is a support function for printing a dominant-distribution network from an adjacency matrix.
}
