% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNonNormalDist.R
\name{SimMixDist}
\alias{SimMixDist}
\title{SimMixDist function}
\usage{
SimMixDist(nInv, mean, std, p1, p2)
}
\arguments{
\item{nInv}{is a number of samples the function will generate.}

\item{mean}{is a mean of a normal distribution part of mixture distribution.}

\item{std}{is a standard deviation of a normal distribution part of mixture distribution.}

\item{p1}{is a ratio of a normal distribution within a mixture distribution.}

\item{p2}{is a ratio of a Cauchy distribution within a mixture distribution.}
}
\value{
This function returns a vector of samples \code{V} generated by a mixture distribution.
}
\description{
SimMixDist is a support function for generating samples from mixture distribution.
The main purpose of this function is to generate samples from non-normal distribution.
}
