% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDOIF.R
\name{EDOIF}
\alias{EDOIF}
\title{Empirical Distribution Ordering Inference Framework (EDOIF)}
\usage{
EDOIF(Values, Group, bootT, alpha, methodType)
}
\arguments{
\item{Values}{is a vector of real-number values}

\item{Group}{is a vector of categories of each real number in Values}

\item{bootT}{is a number of times of sample with replacement for bootstraping.
The default is 1000.
 It must be above zero}

\item{alpha}{is a significance level using in both confidence intervals and ordering inference it has the range [0,1].
The default is 0.05.}

\item{methodType}{is an option for boostrapping methods:either  "perc" or "bca".
The "perc" is the default option.}
}
\value{
this class constructor returns an object of EDOIF class.

\code{obj} an object of EDOIF class that contains the results of ordering inference
that can be print in text mode (print(obj)) or graphic mode (plot(obj)).

The \code{obj} consists of the following variables

\item{Values, Group}{ The main inputs of the framework. They are the double and character vectors respectively. }
\item{bootT, alpha, methodType}{ The number of bootstrapping, significance level, and boostrapping method parameters. }
\item{sortedGroupList}{ The list of names of categories ascendingly ordered by their means. }
\item{sortedmeanList}{ The list of means of categories that are ascendingly ordered. }
\item{MegDiffList[[i]]}{ The mean difference confidence intervals and related information of all categories that have higher means than sortedGroupList[i] category. If methodType = "bca", then MegDiffList[[i]] is an object of dabestr class that can be used to create the Gardner-Altman plot.}
\item{confInvsList[i,]}{ The mean confidence interval of sortedGroupList[i] category. confInvsList[i,1] is a lower bound and confInvsList[i,2] is an upper bound. }
\item{adjMat[i,j]}{ An element of adjacency matrix: one if sortedGroupList[j] category dominates sortedGroupList[i] using Mann-Whitney test, otherwise zero. }
\item{pValMat[i,j]}{ A p-value of Mann-Whitney test for adjMat[i,j]. }
\item{adjDiffMat[i,j]}{ A lower bound of confidence interval of difference mean for sortedGroupList[j] minus sortedGroupList[i] using methodType bootstrap. }
\item{adjBootMat[i,j]}{ One if adjDiffMat[i,j] is positive, otherwise, zero. }
}
\description{
EDOIF is  non-parametric  framework  based on  Esitmation Statistics principle.
Its main purpose is to infer orders of empirical distributions from different categories
base on a probability of finding a value in one distribution that greater than the expectation
of another distribution.
}
\details{
Given a set of ordered-pair of real-category values the framework is capable of
 1) inferring orders of  domination  of  categories  and  representing  orders  in  the form of a graph;
 2) estimating  magnitude  of  difference  between  a  pair  of categories in forms of confidence intervals; and
 3) visualizing  domination  orders  and  magnitudes  of  dif-ference of categories.
}
