% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphOperation.R
\name{getiGraphNetDen}
\alias{getiGraphNetDen}
\title{getiGraphNetDen function}
\usage{
getiGraphNetDen(g)
}
\arguments{
\item{g}{is an object of iGraph class of a dominant-distribution network.}
}
\value{
This function returns a value of network density of of a dominant-distribution network for a given object g.
}
\description{
getiGraphNetDen is a support function for calculating a network density of a dominant-distribution network.
}
