% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotText.R
\name{print.EDOIF}
\alias{print.EDOIF}
\title{print.EDOIF function}
\usage{
\method{print}{EDOIF}(obj)
}
\arguments{
\item{obj}{is an object of EDOIF class that contains the results of ordering inference.}
}
\description{
print.EDOIF is a support function for printing results of orderining inference in text.
}
